requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.0.0
matplotlib>=3.3.0
numpy>=1.19.0
tqdm>=4.50.0
scikit-image>=0.18.0
scipy>=1.7.0 